+++ zc-web-spring-boot-starter/src/main/java/com/zhengcheng/web/WebAutoConfiguration.java
@@ -1,10 +1,14 @@
 package com.zhengcheng.web;
 
 import com.zhengcheng.web.aspect.ControllerLogAspect;
+import com.zhengcheng.web.interceptor.TraceIdInterceptor;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
+import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
@@ -23,6 +27,11 @@ public class WebAutoConfiguration implements WebMvcConfigurer {
     public WebAutoConfiguration() {
     }
 
+    @Bean
+    public HandlerInterceptor getTraceIdInterceptor() {
+        return new TraceIdInterceptor();
+    }
+
     @Override
     public void addResourceHandlers(ResourceHandlerRegistry registry) {
         registry.addResourceHandler("/static/**").addResourceLocations("classpath:/static/");
@@ -32,4 +41,12 @@ public class WebAutoConfiguration implements WebMvcConfigurer {
                 .addResourceLocations("classpath:/META-INF/resources/webjars/");
     }
 
+    //添加拦截
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(getTraceIdInterceptor())//路径拦截器
+                .addPathPatterns("/**")//拦截的请求路径
+                .excludePathPatterns("/error")//排除的请求路径
+                .excludePathPatterns("/static/*");
+    }
 }